Move ChatEntropy to common package
